services/linkedin_service.py
import requests
from typing import List, Optional
from models.linkedin_profile import LinkedInProfile
from bs4 import BeautifulSoup
from config import Config
from services.profile_storage import ProfileStorage

class LinkedInService:
    def __init__(self, cookies: str):
        self.cookies = self._parse_cookies(cookies)
        self.session = requests.Session()
        self.session.cookies.update(self.cookies)
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }
        self.storage = ProfileStorage(Config.PROFILES_DIR)

    def _parse_cookies(self, cookies_str: str) -> dict:
        return dict(cookie.split('=', 1) for cookie in cookies_str.split('; '))

    def search_by_phone(self, phone: str, area_code: str) -> List[LinkedInProfile]:
        # First try to load from storage
        profile = self.storage.load_profile(phone)
        if profile:
            return [profile]
            
        # TODO: Implement actual LinkedIn search
        # This would involve:
        # 1. Using the LinkedIn Search API or scraping search results
        # 2. Filtering by location using area code
        # 3. Creating LinkedInProfile objects from results
        # 4. Saving found profile using self.storage.save_profile()
        return []

    def get_profile_by_url(self, profile_url: str) -> Optional[LinkedInProfile]:
        try:
            response = self.session.get(profile_url, headers=self.headers)
            if response.status_code == 200:
                return self._parse_profile(response.text)
        except Exception as e:
            print(f"Error fetching profile: {e}")
        return None

    def _parse_profile(self, html_content: str) -> Optional[LinkedInProfile]:
        soup = BeautifulSoup(html_content, 'html.parser')
        # TODO: Implement actual profile parsing
        # This would involve extracting:
        # - Name
        # - Location
        # - Current title
        # - Company
        # - Full profile text
        return None

    def parse_profile_text(self, html_content: str) -> str:
        # Parse LinkedIn profile HTML to extract relevant text
        pass