requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
twilio==8.11.0
websockets==12.0
certifi==2023.11.17
python-multipart==0.0.6
beautifulsoup4==4.12.3
python-dotenv==1.0.0
requests==2.32.3